requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma>=0.1.4
google-generativeai>=0.5.0
chromadb
python-dotenv
tiktoken
